refactor: 예상 결제 금액 구하는 유틸 분리
